+++ src/ResourceManager/DataLakeAnalytics/Commands.DataLakeAnalytics.Test/ScenarioTests/AdlaAliasTests.ps1
@@ -375,7 +375,7 @@ function Test-DataLakeAnalyticsAccount
 
 		# Updating Account
 		$tagsToUpdate = @{"TestTag" = "TestUpdate"}
-		$accountUpdated = Set-AdlAnalyticsAccount -ResourceGroupName $resourceGroupName -Name $accountName -Tags $tagsToUpdate
+		$accountUpdated = Set-AdlAnalyticsAccount -ResourceGroupName $resourceGroupName -Name $accountName -Tag $tagsToUpdate
     
 		Assert-AreEqual $accountName $accountUpdated.Name
 		Assert-AreEqual $location $accountUpdated.Location
@@ -714,7 +714,7 @@ function Test-NegativeDataLakeAnalyticsAccount
 
 		# attempt to update a non-existent account
 		$tagsToUpdate = @{"TestTag" = "TestUpdate"}
-		Assert-Throws {Set-AdlAnalyticsAccount -ResourceGroupName $resourceGroupName -Name $fakeaccountName -Tags $tagsToUpdate}
+		Assert-Throws {Set-AdlAnalyticsAccount -ResourceGroupName $resourceGroupName -Name $fakeaccountName -Tag $tagsToUpdate}
 
 		# attempt to get a non-existent account
 		Assert-Throws {Get-AdlAnalyticsAccount -ResourceGroupName $resourceGroupName -Name $fakeaccountName}