+++ src/ResourceManager/Sql/Commands.Sql.Test/ScenarioTests/DatabaseCrudTests.ps1
@@ -563,16 +563,17 @@ function Test-CancelDatabaseOperation
 	.SYNOPSIS
 	Tests listing and cancelling a database operation
 #>
-function Test-CancelDatabaseOperationInternal ($location = "westcentralus")
+function Test-CancelDatabaseOperationInternal
 {
 	# Setup
-	$rg = Create-ResourceGroupForTest
+	$location = Get-Location "Microsoft.Sql" "operations" "Southeast Asia"
+	$rg = Create-ResourceGroupForTest $location
 	$server = Create-ServerForTest $rg $location
 
 	$databaseName = Get-DatabaseName
 	$db = New-AzureRmSqlDatabase -ResourceGroupName $rg.ResourceGroupName -ServerName $server.ServerName -DatabaseName $databaseName `
 		-Edition Standard -MaxSizeBytes 250GB -RequestedServiceObjectiveName S0
-	Assert-AreEqual $db.DatabaseName $databaseName
+	Assert-AreEqual $db.DatabaseName $databaseName 'Create database failed.'
 
 	# Database will be Standard s0 with maxsize: 268435456000 (250GB)
 
@@ -581,9 +582,9 @@ function Test-CancelDatabaseOperationInternal ($location = "westcentralus")
 		# Alter all properties
 		$db1 = Set-AzureRmSqlDatabase -ResourceGroupName $db.ResourceGroupName -ServerName $db.ServerName -DatabaseName $db.DatabaseName `
 			-Edition Standard -RequestedServiceObjectiveName S1
-		Assert-AreEqual $db1.DatabaseName $db.DatabaseName
-		Assert-AreEqual $db1.Edition Standard
-		Assert-AreEqual $db1.CurrentServiceObjectiveName S1
+		Assert-AreEqual $db1.DatabaseName $db.DatabaseName 'Alter db name not equal'
+		Assert-AreEqual $db1.Edition Standard 'Alter db edition not equal'
+		Assert-AreEqual $db1.CurrentServiceObjectiveName S1 'Alter db slo not equal'
 
 		# list and cancel a database operation
 		$dbactivity = Get-AzureRmSqlDatabaseActivity -ResourceGroupName $db.ResourceGroupName -ServerName $db.ServerName -DatabaseName $db.DatabaseName
@@ -598,7 +599,7 @@ function Test-CancelDatabaseOperationInternal ($location = "westcentralus")
 		Catch
 		{
 			$ErrorMessage = $_.Exception.Message
-			Assert-AreEqual True $ErrorMessage.Contains("Cannot cancel database management operation '" + $dbactivityId + "' in the current state")
+			Assert-AreEqual True $ErrorMessage.Contains("Cannot cancel management operation '" + $dbactivityId + "' in the current state")
 		}
 	}
 	finally